GDPR
Using JavaScript and DOM manipulation, simulate the ability for a user to "accept cookie usage". Provide a footer banner where a user can click a link to "Accept Cookies." When the user clicks "Accept Cookies", change the content on the banner to display new content in the footer stating: "Cookies were accepted. Would you like to revoke?" and provide a link for the user to revoke. If the user chooses to revoke, use JavaScript to revert to the original banner content. (In bigger applications, you would need to send the user to a page that describes cookie usage and permissions that that the user can can revoke from there.)